Push calibration 2'nd step. Calibrations are received by master and processed.

+++ app/src/main/java/com/eveningoutpost/dexdrip/GcmActivity.java
@@ -63,7 +63,7 @@ class NewCalibration {
     long timestamp;
     
     @Expose
-    double offset;
+    long offset;
     
     @Expose
     String uuid;
